separo componente, agregos disabled prop, ajusto getRandomFunction
